fix(647): Moved DataFormat, StepEditer component to different path
