+++ packages/ui/src/components/Visualization/Canvas/index.ts
@@ -1,5 +1,5 @@
 export * from './Canvas';
 export * from './CanvasSideBar';
-export * from './DataFormatEditor';
+export * from '../../Form/dataFormat/DataFormatEditor';
 export * from './StepExpressionEditor';
 export * from './canvas.models';
